requirements.txt
agate==1.9.1
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
daff==1.3.46
databricks-sdk==0.33.0
db-dtypes==1.3.0
dbt-adapters==1.7.0
dbt-bigquery==1.8.2
dbt-common==1.10.0
dbt-core==1.8.7
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1
debugpy==1.8.6
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
diff_cover==9.2.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-dataproc==5.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.62.3
gunicorn==23.0.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jinja2-simple-tags==0.6.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
leather==0.4.0
Logbook==1.5.3
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
mashumaro==3.13.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
minimal-snowplow-tracker==0.0.2
more-itertools==10.5.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.35
sqlfluff==3.2.0
sqlfluff-templater-dbt==3.2.0
sqlparse==0.5.1
stack-data==0.6.3
tblib==3.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
prophet
